refactor: introduce server layer
